// bid/src/modules/grpc/proto/v2.pb.ts
export interface GetProductsResponse_Product {
  productId: string;
  sellerId: string;
  productName: string;
  startBid: number;
  sellerName: string;
  productNameAr: string;
  commission: number;
  shipping: number;
  availablePayment: string;
  isExpired: boolean;
  vat: number;
  productImg: string;
  isDeleted: boolean;
  isSold: boolean;
  sellPrice: number;
  sellerCity: string;
  vatPercentage: number;
  sellDate?: string | undefined;
  attributes: Attribute[];
}
// ...
export interface GetMarketPriceByVariantIdResponse {
  minPrice: number;
  maxPrice: number;
}
// ...
export interface BreakDownResponse {
  withPromo: CommissionSummaryResponse | undefined;
  withoutPromo: CommissionSummaryResponse | undefined;
}

export interface CommissionAnalysis {
  commissionTotalPercentage: number;
  commissionTotalFixed: number;
  paymentCommissionExtraFees: number;
  paymentCommission: number;
  paymentCommissionVat: number;
  nonPaymentCommission: number;
  nonPaymentCommissionVat: number;
  paymentCommissionWithVat: number;
  nonPaymentCommissionWithVat: number;
}

export interface CommissionSummaryResponse {
  id: string;
  commission: number;
  commissionVat: number;
  deliveryFee: number;
  deliveryFeeVat: number;
  totalVat: number;
  discount: number;
  grandTotal: number;
  commissionDiscount: number;
  sellPrice: number;
  commissionAnalysis: CommissionAnalysis | undefined;
  paymentId: string;
}
// ...
export interface Title {
  arName: string;
  enName: string;
}

export interface Value {
  arName: string;
  enName: string;
}

export interface Attribute {
  title: Title | undefined;
  value: Value | undefined;
}